refactor wallet
